Implement MNIST loader

+++ data/mnist.py
@@ -44,8 +44,8 @@ def load_mnist(
     test_images = _read_idx_images(data_dir / "t10k-images.idx3-ubyte")
     test_labels = _read_idx_labels(data_dir / "t10k-labels.idx1-ubyte")
 
-    X_train_full = train_images.reshape(-1, 784).astype(np.float32) / 255.0
-    X_test = test_images.reshape(-1, 784).astype(np.float32) / 255.0
+    X_train_full = train_images.reshape(-1, 784).astype(np.float64) / 255.0
+    X_test = test_images.reshape(-1, 784).astype(np.float64) / 255.0
 
     rng = np.random.default_rng(seed)
     perm = rng.permutation(X_train_full.shape[0])
